moving Stories.execute_line, and Stories.execute_block into Lexicon
